requirements.txt
coolname==1.1.0
matplotlib==3.1.1
numpy==1.17.2
scikit-learn==0.21.3
tensorboard==2.0.0
tensorflow-gpu==2.0.0
tensorflow-probability==0.8.0
